+++ lib/services/index.ts
@@ -14,4 +14,4 @@
 export * from './AutoRenewService';
 export * from './SyncStorageService';
 export * from './LeaderElectionService';
-export * from './InactiveTabService';
+export * from './RenewOnTabActivationService';